requirements.txt
pyspark==3.5.1
dbt-core==1.7.5
dbt-snowflake==1.7.5
protobuf==4.23.4
pandas==2.3.2
requests==2.32.5
pyarrow==21.0.0
reportlab==4.4.3